# docs/swagger/swagger.yaml
definitions:
...
  handlers.eventResponse:
    properties:
      created_at:
        type: string
      creator_id:
        type: integer
      description:
        type: string
      end_time:
        type: string
      id:
        type: integer
      location:
        type: string
      meeting_url:
        type: string
      start_time:
        type: string
      status:
        type: string
      title:
        type: string
    type: object
  handlers.loginRequest:
    properties:
      email:
        example: petr@example.com
        type: string
      password:
        example: super-secret
        type: string
    type: object
  handlers.loginResponse:
    properties:
      email:
        type: string
      id:
        type: integer
      role:
        type: string
      token:
        type: string
    type: object
  handlers.meResponse:
    properties:
      email:
        type: string
      id:
        type: integer
      role:
        type: string
    type: object
...
  handlers.registerRequest:
    properties:
      email:
        example: petr@example.com
        type: string
      password:
        example: super-secret
        type: string
      role:
        enum:
        - admin
        - organizer
        - employee
        example: employee
        type: string
    type: object
  handlers.registerResponse:
    properties:
      email:
        type: string
      id:
        type: integer
      role:
        type: string
    type: object
  handlers.taskResponse:
    properties:
      assignee_id:
        type: integer
      created_at:
        type: string
      description:
        type: string
      due_date:
        type: string
      event_id:
        type: integer
      id:
        type: integer
      status:
        type: string
      title:
        type: string
    type: object
...
  handlers.userResponse:
    properties:
      email:
        type: string
      id:
        type: integer
      role:
        type: string
    type: object
info:
  contact: {}
  description: Сервис для управления рабочими мероприятиями.
  title: Meeting Service API
  version: "1.0"
paths:
...
  /login:
    post:
      consumes:
      - application/json
      parameters:
      - description: Email и пароль
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/handlers.loginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.loginResponse'
        "401":
          description: invalid email or password
          schema:
            type: string
      summary: Вход в систему, возвращает JWT
      tags:
      - auth
  /me:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.meResponse'
        "401":
          description: unauthorized
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Профиль текущего пользователя
      tags:
      - auth
  /me/events:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handlers.eventResponse'
            type: array
        "401":
          description: unauthorized
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Встречи, в которых я создатель или участник
      tags:
      - auth
  /me/tasks:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handlers.taskResponse'
            type: array
        "401":
          description: unauthorized
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Задачи, назначенные на меня
      tags:
      - auth
  /register:
    post:
      consumes:
      - application/json
      parameters:
      - description: Учётные данные
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/handlers.registerRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handlers.registerResponse'
        "400":
          description: validation error
          schema:
            type: string
      summary: Регистрация нового пользователя
      tags:
      - auth
  /users:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handlers.userResponse'
            type: array
        "401":
          description: unauthorized
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Список всех пользователей
      tags:
      - users
swagger: "2.0"